backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
sqlparse==0.5.3
qdrant-client==1.12.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
